+++ CHANGELOG.md
@@ -5,10 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[0.3.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/johnarp/proficiency-tracker/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/johnarp/proficiency-tracker/compare/v0.0.1...v0.1.0
 [0.0.1]: https://github.com/johnarp/proficiency-tracker/releases/tag/v0.0.1
+
+## [0.3.0] - 2026-03-15
+
+### Added
+
+- Proficiency icons
+- Toggle to show hero names
+- Hero card size options
+- About section in settings
+
+### Changed
+
+- Improved readability for customize and settings
+- Scrollbar disappearing does not move the settings icon
 
 ## [0.2.0] - 2026-03-14